requirements.txt
streamlit==1.31.0
anthropic==0.18.1
openai==1.12.0
google-generativeai==0.3.2
python-dotenv==1.0.0
supabase==2.3.4
pandas==2.2.0
numpy==1.26.3
requests==2.31.0
Pillow==10.2.0
